+++ packages/mongoose/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.1](https://github.com/Oneflow/davinci/compare/@davinci/mongoose@0.8.0...@davinci/mongoose@0.8.1) (2019-11-01)
+
+
+### Bug Fixes
+
+* **validation:** safer check for validationOptions ([b51ced2](https://github.com/Oneflow/davinci/commit/b51ced2))
+
+
+
+
+
+# [0.8.0](https://github.com/Oneflow/davinci/compare/@davinci/mongoose@0.7.2...@davinci/mongoose@0.8.0) (2019-11-01)
+
+
+### Features
+
+* **partialValidation:** added partial validation for schemas in route decorators ([d485871](https://github.com/Oneflow/davinci/commit/d485871))
+
+
+
+
+
 ## [0.7.2](https://github.com/Oneflow/davinci/compare/@davinci/mongoose@0.7.1...@davinci/mongoose@0.7.2) (2019-10-31)
 
 **Note:** Version bump only for package @davinci/mongoose